ref(rcs): move command handling to spot-tv rcs
